api: split stock api
分离出获取股票的函数，封装成单独的类专用于获取网易股市。
解耦之后，日后如有需要可以支持多个获取股票状态的源
